+++ CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG for Sulu
     * BUGFIX      #2915 [MediaBundle]         Fixed height of badges in media-selection
     * ENHANCEMENT #2860 [ContentBundle]       Added url to internal links and smart-content
     * FEATURE     #2914 [CategoryBundle]      Added description and medias to category
+    * FEATURE     #2908 [MediaBundle]         Created cropping-slide for media-edit-overlay
     * FEATURE     #2877 [MediaBundle]         Rest-Api for image-formats and format options
     * ENHANCEMENT #2910 [ContentBundle]       Updated husky and added placeholder param to date
     * ENHANCEMENT #2912 [CategoryBundle]      Implemented category-bundle entities extensible
@@ -12,6 +13,7 @@ CHANGELOG for Sulu
     * ENHANCEMENT #2903 [RouteBundle]         Get class mapping configuration by class name or inheritance chain
     * ENHANCEMENT #2904 [RouteBundle]         Added route-provider cache
     * BUGFIX      #2893 [ContentBundle]       Fixed stop overlay-component for teaser-selection
+    * FEATURE     #2891 [MediaBundle]         Created area-selection frontend component in media-bundle
     * BUGFIX      #2878 [ContentBundle]       Added instanceof check for shadow-behavior functions
     * FEATURE     #2875 [RouteBundle]         Added route-content-type
     * BUGFIX      #2513 [PersistanceBundle]   Fix doctrine generator commands